gs1-image-checker/requirements.txt
streamlit
opencv-python
numpy
pillow
pytesseract
requests
